refactor: move esbuild index generator, code bundle option and execution results
This commit extracts code in separate files.
