+++ packages/angular_devkit/build_angular/src/tools/esbuild/application-code-bundle.ts
@@ -1,0 +1,140 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+import type { BuildOptions } from 'esbuild';
+import { NormalizedBrowserOptions } from '../../builders/browser-esbuild/options';
+import { SourceFileCache, createCompilerPlugin } from '../../tools/esbuild/angular/compiler-plugin';
+import { createExternalPackagesPlugin } from '../../tools/esbuild/external-packages-plugin';
+import { createSourcemapIngorelistPlugin } from '../../tools/esbuild/sourcemap-ignorelist-plugin';
+import { getFeatureSupport } from '../../tools/esbuild/utils';
+import { createVirtualModulePlugin } from '../../tools/esbuild/virtual-module-plugin';
+
+export function createCodeBundleOptions(
+  options: NormalizedBrowserOptions,
+  target: string[],
+  browsers: string[],
+  sourceFileCache?: SourceFileCache,
+): BuildOptions {
+  const {
+    workspaceRoot,
+    entryPoints,
+    optimizationOptions,
+    sourcemapOptions,
+    tsconfig,
+    outputNames,
+    outExtension,
+    fileReplacements,
+    externalDependencies,
+    preserveSymlinks,
+    stylePreprocessorOptions,
+    advancedOptimizations,
+    inlineStyleLanguage,
+    jit,
+    tailwindConfiguration,
+  } = options;
+
+  const buildOptions: BuildOptions = {
+    absWorkingDir: workspaceRoot,
+    bundle: true,
+    format: 'esm',
+    entryPoints,
+    entryNames: outputNames.bundles,
+    assetNames: outputNames.media,
+    target,
+    supported: getFeatureSupport(target),
+    mainFields: ['es2020', 'browser', 'module', 'main'],
+    conditions: ['es2020', 'es2015', 'module'],
+    resolveExtensions: ['.ts', '.tsx', '.mjs', '.js'],
+    metafile: true,
+    legalComments: options.extractLicenses ? 'none' : 'eof',
+    logLevel: options.verbose ? 'debug' : 'silent',
+    minify: optimizationOptions.scripts,
+    pure: ['forwardRef'],
+    outdir: workspaceRoot,
+    outExtension: outExtension ? { '.js': `.${outExtension}` } : undefined,
+    sourcemap: sourcemapOptions.scripts && (sourcemapOptions.hidden ? 'external' : true),
+    splitting: true,
+    tsconfig,
+    external: externalDependencies,
+    write: false,
+    platform: 'browser',
+    preserveSymlinks,
+    plugins: [
+      createSourcemapIngorelistPlugin(),
+      createCompilerPlugin(
+        // JS/TS options
+        {
+          sourcemap: !!sourcemapOptions.scripts,
+          thirdPartySourcemaps: sourcemapOptions.vendor,
+          tsconfig,
+          jit,
+          advancedOptimizations,
+          fileReplacements,
+          sourceFileCache,
+          loadResultCache: sourceFileCache?.loadResultCache,
+        },
+        // Component stylesheet options
+        {
+          workspaceRoot,
+          optimization: !!optimizationOptions.styles.minify,
+          sourcemap:
+            // Hidden component stylesheet sourcemaps are inaccessible which is effectively
+            // the same as being disabled. Disabling has the advantage of avoiding the overhead
+            // of sourcemap processing.
+            !!sourcemapOptions.styles && (sourcemapOptions.hidden ? false : 'inline'),
+          outputNames,
+          includePaths: stylePreprocessorOptions?.includePaths,
+          externalDependencies,
+          target,
+          inlineStyleLanguage,
+          preserveSymlinks,
+          browsers,
+          tailwindConfiguration,
+        },
+      ),
+    ],
+    define: {
+      // Only set to false when script optimizations are enabled. It should not be set to true because
+      // Angular turns `ngDevMode` into an object for development debugging purposes when not defined
+      // which a constant true value would break.
+      ...(optimizationOptions.scripts ? { 'ngDevMode': 'false' } : undefined),
+      'ngJitMode': jit ? 'true' : 'false',
+    },
+  };
+
+  if (options.externalPackages) {
+    buildOptions.plugins ??= [];
+    buildOptions.plugins.push(createExternalPackagesPlugin());
+  }
+
+  const polyfills = options.polyfills ? [...options.polyfills] : [];
+  if (jit) {
+    polyfills.push('@angular/compiler');
+  }
+
+  if (polyfills?.length) {
+    const namespace = 'angular:polyfills';
+    buildOptions.entryPoints = {
+      ...buildOptions.entryPoints,
+      ['polyfills']: namespace,
+    };
+
+    buildOptions.plugins?.unshift(
+      createVirtualModulePlugin({
+        namespace,
+        loadContent: () => ({
+          contents: polyfills.map((file) => `import '${file.replace(/\\/g, '/')}';`).join('\n'),
+          loader: 'js',
+          resolveDir: workspaceRoot,
+        }),
+      }),
+    );
+  }
+
+  return buildOptions;
+}
